requirements.txt
asgiref==3.4.1
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
distlib==0.3.3
Django==3.2.9
djangorestframework==3.12.4
filelock==3.4.0
platformdirs==2.4.0
pycodestyle==2.8.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
virtualenv==20.10.0
